requirements.txt
fastapi
uvicorn
sqlalchemy
requests
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
numpy
scikit-learn
better_profanity
fastapi-mail
